+++ packages/ui-components/web/src/Icons/index.ts
@@ -1,5 +1,8 @@
 // @create-index
 
+export { default as InfoIcon } from './InfoIcon';
+
+// plugin-commons
 // export { default as AlignCenterIcon } from './AlignCenterIcon';
 // export { default as BackArrowIcon } from './BackArrowIcon';
 // export { default as HeightIcon } from './HeightIcon';
@@ -26,3 +29,4 @@ export { default as MoreIcon } from './MoreIcon';
 export { default as ClearIcon } from './ClearIcon';
 export { default as SearchIcon } from './SearchIcon';
 export { default as DropdownArrowIcon } from './DropdownArrowIcon';
+export { default as CheckIcon } from './CheckIcon';