+++ packages/ui-components/web/src/index.ts
@@ -27,5 +27,8 @@ export { default as MediaItemErrorMsg } from './Components/MediaItemErrorMsg';
 export { default as TextSearchInput } from './Components/TextSearchInput';
 export { default as Separator } from './Components/Separator';
 export { default as Dropdown } from './Components/Dropdown';
+export { default as InfoIcon } from './Components/InfoIcon';
+export { default as Checkbox } from './Components/Checkbox';
+export { default as RadioGroup } from './Components/RadioGroup';
 
 // export { getSelectionStyles } from './Utils/inlineStyleUtils';