+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - `editor-common`
   - [#2622](https://github.com/wix-incubator/rich-content/pull/2622) moved textSearchInput/Separator components into ui-components folder
 
+- `editor-common`
+  - [#2618](https://github.com/wix-incubator/rich-content/pull/2618) moved RadioGroup/InfoIcon/Checkbox components into ui-components folder
 - `ricos-content`
   - [#2610](https://github.com/wix/ricos/pull/2610) tiptap converters fixed
   - [#2628](https://github.com/wix/ricos/pull/2628) `extract` & `modify` API exposed via libs folder